requirements.txt
-i https://pypi.org/simple

asgiref==3.6.0 ; python_version >= '3.7'

async-timeout==4.0.2 ; python_version >= '3.6'

certifi==2022.12.7 ; python_version >= '3.6'

cffi==1.15.1

charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'

crispy-bootstrap5==0.7

cryptography==38.0.4

defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

django==4.1.4

django-admin-interface==0.24.2

django-allauth==0.51.0

django-appconf==1.0.5 ; python_version >= '3.6'

django-axes==5.40.1

django-ckeditor-5==0.2.2

django-colorfield==0.8.0

django-compressor==4.1

django-crispy-forms==1.14.0

django-db-logger==0.1.12

django-etc==1.4.0

django-filter==22.1

django-hitcount==1.3.5

django-ipware==4.0.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

django-phonenumber-field==7.0.1

django-ratelimit==4.0.0

gunicorn==20.1.0

idna==3.4 ; python_version >= '3.5'

oauthlib==3.2.2 ; python_version >= '3.6'

phonenumbers==8.13.3

pillow==9.3.0

pycparser==2.21

pyjwt[crypto]==2.6.0 ; python_version >= '3.7'

python-dotenv==0.21.0

python3-openid==3.2.0

rcssmin==1.1.0

redis==4.4.0

requests==2.28.1 ; python_version >= '3.7' and python_version < '4'

requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

rjsmin==1.2.0

setuptools==65.6.3 ; python_version >= '3.7'

six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.4.3 ; python_version >= '3.5'

tzdata==2022.7 ; sys_platform == 'win32'

urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

whitenoise==6.2.0

